Update README.md
ALGORITHM:

(1) A list 'answer' is created which contains all the letters of the word to be guessed.
(2) An empty list 'display[]' is created which is then extended to contain all the elements of the list answer.
(3) Another empty list 'used[]' is created which is extended with the list display to contain all its elements. But this list is modified each time a letter is guessed. All the guesssed letters are removed and the list only contains the letters which are still to be guessed by the player. 
(4) A 'for' loop is used to replace each element in the list with '_' and a space is put between each '_'(using join command) to distinguish the number of dashes.
(5) A counter is used to keep asking the player to guess the letter until all letters have been guessed.
(6) Another 'for' loop is used to traverse through the list answer.
(7) An 'if' statement is used to check whether the letter guessed by the player is in the list 'answer' and also checks whether that letter is present in the list 'used'.
(8) The dash in the list-display is then replaced by the actual letter in the list.
(9)  Remove the guessed letter from the list 'used'.
(10) Print the number of guessed letters and the number of letters still to be guessed.
(11) Once all the letters have been guessed, print the new string with a blankspace between all the letters.Finally, display a message that the word has been guessed.